refactor: Split painter script
